+++ CHANGELOG.md
@@ -130,6 +130,7 @@
 - Run maintenance daily and generate stats in new stats script and added them to seed script !2379 @chriswalg
 - Moved variable DELAY_MICRO_SECONDS_BETWEEN_MAILS to config file #1389 !2383 @chriswalg
 - Replace `node-sass` with `sass` #1047 !2404 @schanso
+- Added mayRole function to Session class and made may function deprecated #1277 !2341 @alex.simm
 
 ## Development documetation
 - Add generator for database documentation !2275 @KoeMai